+++ src/lib/enums/binary-resources.ts
@@ -1,3 +1,5 @@
+import { getAllEnumValues } from "../common/helpers";
+
 /**
  * Types for binary resources (i.e. any resource that follows a BT). Note that
  * this enum is intentionally incomplete, as other binary resources are not yet
@@ -9,6 +11,15 @@ enum BinaryResourceType {
   DstImage = 0x00B2D882,
   SimData = 0x545AC67A,
   StringTable = 0x220557DA,
+}
+
+namespace BinaryResourceType {
+  /**
+   * Returns an array of all binary resource types.
+   */
+  export function all(): BinaryResourceType[] {
+    return getAllEnumValues(BinaryResourceType);
+  }
 }
 
 // `export default enum` not supported by TS