+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add `getLocaleCode()` to StringTableLocale.
 - Add `indexResources()` to Package.
+- Add optional `key` attribute to ResourcePosition.
 
 ## [0.4.2] - 2022/06/04
 ### Added